Added the following methods: checkForConflicts() - This loops through all given courses and increments an integer up everytime it detects a conflict.
isConflicting() - This is a boolean method that returns true if any start and/or end times are found to hold equal values.

printConflictDetails() - Just prints the courses that are conflicting, and shows the conflict details (times, day, and subject)

conCurrentCourses() - Creates a schedule via an array list displaying all courses that can be registered in with no conflicts for a given semester.

scheduleHasConflicts() - Checks to see if 2 or more courses within a given schedule are conflicting

printConcurrentCourses() - Prints a list displaying the schedule that can be taken during a specific semester